+++ semantic_conventions/lib/opentelemetry/semconv/incubating/db/attributes.rb
@@ -326,7 +326,12 @@ module OpenTelemetry
         # up with the parameterized placeholders present in `db.query.text`.
         #
         # It is RECOMMENDED to capture the value as provided by the application
-        # without attempting to do any case normalization.
+        # without attempting to do any case normalization or sanitization.
+        #
+        # Instrumentations SHOULD NOT capture `db.query.parameter.<key>` by default
+        # since values may contain PII or sensitive details.
+        # Application operators are expected to enable specific keys depending
+        # on their privacy and security considerations.
         #
         # `db.query.parameter.<key>` SHOULD NOT be captured on batch operations.
         #
@@ -376,7 +381,7 @@ module OpenTelemetry
         # Deprecated, use `db.namespace` instead.
         #
         # @note Stability Level: development
-        # @deprecated Uncategorized.
+        # @deprecated Replaced by `db.namespace` (string).
         DB_REDIS_DATABASE_INDEX = 'db.redis.database_index'
 
         # Number of rows returned by the operation.