+++ metrics_api/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Release History: opentelemetry-metrics-api
 
+### v0.7.0 / 2026-08-04
+
+* ADDED: Add attributes to Meter InstrumentationScope (#2273)
+
 ### v0.6.1 / 2026-07-08
 
 * DOCS: Update source for gemspec links to RubyDoc (#2123)